+++ src/javascript/binary/base/page.js
@@ -693,18 +693,19 @@ Header.prototype = {
     show_or_hide_login_form: function() {
         if (!this.user.is_logged_in || !this.client.is_logged_in) return;
         var all_accounts = $('#all-accounts');
-        var that = this;
-        $('.nav-menu').unbind('click').on('click', function(event){
+        function openCloseMenu(event) {
             event.stopPropagation();
             if (all_accounts.css('opacity') == 1 ) {
-                that.animate_disappear(all_accounts);
+                page.header.animate_disappear(all_accounts);
             } else {
-                that.animate_appear(all_accounts);
+                page.header.animate_appear(all_accounts);
             }
-        });
-        $(document).unbind('click'). on('click', function(){
-            that.animate_disappear(all_accounts);
-        });
+        }
+        function closeMenu() {
+            page.header.animate_disappear(all_accounts);
+        }
+        $('.nav-menu').unbind('click', openCloseMenu).on('click', openCloseMenu);
+        $(document).unbind('click', closeMenu).on('click', closeMenu);
         var loginid_select = '';
         var loginid_array = this.user.loginid_array;
         for (var i=0; i < loginid_array.length; i++) {
