+++ src/javascript/binary/websocket_pages/mb_trade/mb_notifications.js
@@ -1,3 +1,5 @@
+var MBPrice = require('./mb_price').MBPrice;
+
 /*
  * Notifications manages various notification messages
  *
