+++ src/javascript/binary/websocket_pages/trade/analysis.js
@@ -1,3 +1,5 @@
+var MBTradePage = require('../mb_trade/mb_tradepage').MBTradePage;
+
 /*
  * This file contains the code related to loading of trading page bottom analysis
  * content. It will contain jquery so as to compatible with old code and less rewrite
