+++ src/javascript/binary/websocket_pages/user/account/portfolio/portfolio.init.js
@@ -4,6 +4,7 @@ var format_money = require('../../../../common_functions/currency_to_symbol').fo
 var buildOauthApps = require('../../../../common_functions/get_app_details').buildOauthApps;
 var addTooltip = require('../../../../common_functions/get_app_details').addTooltip;
 var showTooltip = require('../../../../common_functions/get_app_details').showTooltip;
+var MBTradePage = require('../../../../websocket_pages/mb_trade/mb_tradepage').MBTradePage;
 
 var PortfolioWS =  (function() {
     'use strict';
